Refactor raid constants
